feat: Remove stars from string
Added a new file for a C++ program that removes all the stars from a given string. This program takes user input and uses a loop to check each character in the string. If the character is a star, it is removed from the string. The resulting string is then output to the user. This program is useful for cleaning up input that may contain unwanted characters, such as when parsing user input for a search function.