feat: Add cpp/2218.maximum-value-of-k-coins-from-piles.cpp
This commit adds a new file cpp/2218.maximum-value-of-k-coins-from-piles.cpp which contains a solution to the problem of finding the maximum value of k coins that can be obtained from piles of coins of varying sizes. The solution involves sorting the piles in descending order and then taking the maximum k piles. The time complexity of the solution is O(n log n) due to the sorting step.