Add new solutions to Online Majority Element in Subarray and Number of Ways to Form a Target String Given a Dictionary
Added two new files containing solutions for the Online Majority Element in Subarray and Number of Ways to Form a Target String Given a Dictionary problems. These solutions use efficient algorithms to solve the problems and have been thoroughly tested for accuracy and performance. The solutions are written in C++ and are intended for use by other developers who are working on similar problems.