feat: Add solution for problem 2465 - Number of Distinct Averages
This commit adds a new file, cpp/2465.number-of-distinct-averages.cpp,
which contains the solution for the problem 2465 - Number of Distinct
Averages. The solution uses a combination of math and dynamic
programming to compute the number of distinct averages that can be
formed from a given set of integers. The time complexity of the
algorithm is O(n^2), where n is the size of the input set.