feat: Add new solutions and modify existing ones
This commit adds a new solution for problem #1751, 'Maximum Number of Events That Can be Attended II', and a new solution for problem #1888, 'Minimum Number of Flips to Make the Binary String Alternating'. Additionally, the solution for problem #18, '4 Sum', has been modified. There are no breaking changes or major refactorings in this commit.