Add new solutions for LeetCode problems
Added solutions for 6 LeetCode problems: 1434, 1499, 1942, 1943, 1947, and 715. These solutions are written in C++ and include the necessary files to run them. The 1434 and 1499 problems involve finding the number of ways to wear different hats and maximizing an equation, respectively. The 1942 and 1943 problems involve finding the smallest unoccupied chair and describing a painting, respectively. The 1947 problem involves finding the maximum compatibility score sum, and the 715 problem involves implementing a range